+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/editors/electron-add-link-context-menu.spec.ts
@@ -10,160 +10,11 @@
  * 6. The wikilink creates an edge in the Cytoscape graph
  */
 
-import { test as base, expect, _electron as electron } from '@playwright/test';
-import type { ElectronApplication, Page } from '@playwright/test';
+import { expect } from '@playwright/test';
 import * as path from 'path';
-import * as fs from 'fs/promises';
-import * as os from 'os';
-import type { Core as CytoscapeCore, NodeSingular } from 'cytoscape';
-import type { EditorView } from '@codemirror/view';
-import type { ElectronAPI } from '@/shell/electron';
-
-const PROJECT_ROOT = path.resolve(process.cwd());
-
-interface ExtendedWindow {
-  cytoscapeInstance?: CytoscapeCore;
-  electronAPI?: ElectronAPI;
-}
-
-interface CodeMirrorElement extends HTMLElement {
-  cmView?: { view: EditorView };
-}
-
-const test = base.extend<{
-  electronApp: ElectronApplication;
-  appWindow: Page;
-  testVaultPath: string;
-}>({
-  electronApp: async ({}, use, testInfo) => {
-    // Create temp userData directory
-    const tempUserDataPath = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-add-link-context-menu-'));
-
-    // Create the watched folder (what config points to)
-    const watchedFolder = path.join(tempUserDataPath, 'test-vault');
-    await fs.mkdir(watchedFolder, { recursive: true });
-
-    // Create the actual vault path with default suffix 'voicetree'
-    const projectRoot = path.join(watchedFolder, 'voicetree');
-    await fs.mkdir(projectRoot, { recursive: true });
-
-    // Create the main test node (source node for the link)
-    const sourceContent = `---
-position:
-  x: 100
-  y: 100
----
-# Source Node
-
-This is the source node where we will add a link.
-
-Some content here.
-`;
-    await fs.writeFile(path.join(projectRoot, 'source-node.md'), sourceContent, 'utf-8');
-
-    // Create a target node that we will link to
-    const targetContent = `---
-position:
-  x: 300
-  y: 100
----
-# Target Node
-
-This is the target node we will link to.
-`;
-    await fs.writeFile(path.join(projectRoot, 'target-node.md'), targetContent, 'utf-8');
-
-    // Create another node to verify autocomplete filtering
-    const anotherContent = `---
-position:
-  x: 200
-  y: 200
----
-# Another Node
-
-This is another node in the graph.
-`;
-    await fs.writeFile(path.join(projectRoot, 'another-node.md'), anotherContent, 'utf-8');
-
-    // Write config to auto-load the vault
-    const configPath = path.join(tempUserDataPath, 'voicetree-config.json');
-    await fs.writeFile(configPath, JSON.stringify({ lastDirectory: watchedFolder }, null, 2), 'utf8');
-    console.log('[Test] Watched folder:', watchedFolder);
-    console.log('[Test] Vault path (with suffix):', projectRoot);
-
-    // Store projectRoot for test access
-    (testInfo as unknown as { projectRoot: string }).projectRoot = projectRoot;
-
-    const electronApp = await electron.launch({
-      args: [
-        path.join(PROJECT_ROOT, 'dist-electron/main/index.js'),
-        `--user-data-dir=${tempUserDataPath}`
-      ],
-      env: {
-        ...process.env,
-        NODE_ENV: 'test',
-        HEADLESS_TEST: '1',
-        MINIMIZE_TEST: '1',
-        VOICETREE_PERSIST_STATE: '1'
-      },
-      timeout: 30000
-    });
-
-    await use(electronApp);
-
-    // Graceful shutdown
-    try {
-      const window = await electronApp.firstWindow();
-      await window.evaluate(async () => {
-        const api = (window as unknown as ExtendedWindow).electronAPI;
-        if (api) {
-          await api.main.stopFileWatching();
-        }
-      });
-      await window.waitForTimeout(300);
-    } catch {
-      console.log('[Test] Could not stop file watching during cleanup');
-    }
-
-    await electronApp.close();
-    console.log('[Test] Electron app closed');
-
-    // Cleanup temp directory
-    await fs.rm(tempUserDataPath, { recursive: true, force: true });
-    console.log('[Test] Cleaned up temp directory');
-  },
-
-  testVaultPath: async ({}, use, testInfo) => {
-    await use((testInfo as unknown as { projectRoot: string }).projectRoot);
-  },
-
-  appWindow: async ({ electronApp }, use) => {
-    const window = await electronApp.firstWindow();
-
-    window.on('console', msg => {
-      console.log(`BROWSER [${msg.type()}]:`, msg.text());
-    });
-
-    window.on('pageerror', error => {
-      console.error('PAGE ERROR:', error.message);
-    });
-
-    await window.waitForLoadState('domcontentloaded', { timeout: 60000 });
-    await window.waitForFunction(() => (window as unknown as ExtendedWindow).cytoscapeInstance, { timeout: 20000 });
-
-    // Wait for auto-load to complete by polling for cytoscape nodes
-    // The vault has 3 files, so wait for at least 1 node to load
-    await window.waitForFunction(() => {
-      const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-      if (!cy) return false;
-      return cy.nodes().length >= 1;
-    }, { timeout: 15000 });
-
-    await window.waitForTimeout(500);
-
-    await use(window);
-  }
-});
+import type { NodeSingular } from 'cytoscape';
+import { PROJECT_ROOT, test } from './electron-add-link-context-menu/fixtures';
+import type { CodeMirrorElement, ExtendedWindow } from './electron-add-link-context-menu/fixtures';
 
 test.describe('Add Link Context Menu', () => {
   test('should show context menu with Add Link option on right-click', async ({ appWindow }) => {
